SoundEmu_Start: allow user-specific device lists
This allows client applications to supply additional sound devices.
(Or just additional sound cores for an existing device.)
The "player" APIs are currently unaffected of this and will get the
option at a later point.

This commit also introduces a new structure DEV_DECL that keeps basic
information about a sound chip.